Keeping codepoint logic together
